ended product

+++ project/Models/Bushel.py
@@ -2,4 +2,4 @@ class Bushel:
     quart_const = 32
 
     def __init__(self, amount):
-        self.amount = float(amount) * self.quart_const
+        self.amount = int(amount) * self.quart_const
